django/requirements.txt
boto==2.38.0
mock==1.0.1
Pillow==3.3.0
Django==1.10.5
django-angular==0.7.16
django-cors-headers==1.1.0
django-extensions==1.7.3
django-filter==0.9.2
django-jsonfield==1.0.0
django-localflavor-us==1.1
django-model-utils==2.5
django-storages-redux==1.3
django-storages==1.4.1
djangorestframework==3.3.3
djangorestframework_jwt==1.8.0
djangorestframework-recursive==0.1.1
elasticsearch==2.3.0
jsonschema==2.5.1
nose==1.3.7
psycopg2==2.6.1
pypdf2==1.26.0
python-dateutil==2.5.3
pytz==2016.4
redis==2.10.5
requests-aws4auth==0.9
requests==2.10.0
responses==0.5.1
six==1.10.0
sorl-thumbnail==12.3
statistics==1.0.3.5
testfixtures==4.10.0
tinys3==0.1.11
unicodecsv==0.14.1
unittest2==1.1.0
validate_email==1.3
django-ordered-model==1.2.1
xmltodict==0.10.2
numpy==1.11.2
Wand==0.4.4
django_environ==0.4.1
fonttools==3.4.0
gunicorn==19.6.0
python3-ghostscript==0.5.0
aiohttp-wsgi==0.6.5
aiohttp==1.2.0
ujson==1.35
raven==5.32.0
opbeat==3.5.2
django_hosts==2.0
whitenoise==2.0.6
TwitterAPI==2.4.4
cleverbot==2.0.0
PbPython==0.1.9
